requirements.txt
autopep8==1.2.4
decorator==4.0.10
diff-match-patch==20121119
Django==1.10.3
django-bcrypt==0.9.2
django-cors-headers==1.2.2
django-crispy-forms==1.6.0
django-extensions==1.6.7
django-filter==0.15.3
django-import-export==0.4.2
django-silk==0.5.7
djangorestframework==3.5.2
djangorestframework-filters==0.9.0
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
facepy==1.0.8
gunicorn==19.6.0
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
nltk==3.2.1
numpy==1.11.2
oauthlib==2.0.1
pandas==0.19.0
pep8==1.7.0
praw==3.6.0
psycopg2==2.6.2
py-bcrypt==0.4
Pygments==2.1.3
PyJWT==1.4.0
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.7.0
scipy==0.18.1
six==1.10.0
sqlparse==0.1.19
tablib==0.11.2
tweepy==3.5.0
tweetpy==0.1
twython==3.4.0
update-checker==0.13
Werkzeug==0.11.11
